+++ SupportTicketApi/SupportTicketApi.Api/Program.cs
@@ -1,5 +1,5 @@
-using SupportTicketApi.Data;
-using SupportTicketApi.Models;
+using SupportTicketApi.Data.Contexts;
+using SupportTicketApi.Data.Models;
 
 var builder = WebApplication.CreateBuilder(args);
 builder.AddSqlServerDbContext<TicketContext>("sqldata");
