+++ SupportTicketApi/SupportTicketApi.AppHost/Program.cs
@@ -4,7 +4,7 @@ var sql = builder.AddSqlServer("sql", port: 14329)
                  .AddDatabase("sqldata")
                  .WithEndpoint(port: 14330);
 
-builder.AddProject<Projects.SupportTicketApi_ApiModel>("api")
+builder.AddProject<Projects.SupportTicketApi_Api>("api")
     .WithReference(sql);
 
 builder.Build().Run();
